pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-genai>=1.25.0",
    "jinja2>=3.1.6",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "trafilatura>=2.0.0",
]
